requirements.txt
python2.x
xmpppy
